# docs/openapi-production.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PRAXIS OS Product API",
    "description": "Authenticated decision-analysis and simulation beta; not certified operational guidance.",
    "version": "0.9.0"
  },
  "paths": {
    "/v1/decisions/analyze": {
      "post": {
        "summary": "Policy Analysis",
        "operationId": "policy_analysis_v1_decisions_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/v1/finance/statements": {
      "post": {
        "summary": "Statements",
        "operationId": "statements_v1_finance_statements_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinanceInputs"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/finance/balance-sheet": {
      "post": {
        "summary": "Balance",
        "operationId": "balance_v1_finance_balance_sheet_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BalanceSheet"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/finance/valuation": {
      "post": {
        "summary": "Valuation",
        "operationId": "valuation_v1_finance_valuation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValuationInputs"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/finance/capital-allocation": {
      "post": {
        "summary": "Allocation",
        "operationId": "allocation_v1_finance_capital_allocation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CapitalAllocation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/decision-models": {
      "get": {
        "summary": "Decisions",
        "operationId": "decisions_v1_decision_models_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "summary": "Create Decision",
        "operationId": "create_decision_v1_decision_models_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Decision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/decision-models/{identifier}": {
      "get": {
        "summary": "Decision",
        "operationId": "decision_v1_decision_models__identifier__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/decision-models/{identifier}/history": {
      "get": {
        "summary": "History",
        "operationId": "history_v1_decision_models__identifier__history_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/decision-models/{identifier}/feedback": {
      "post": {
        "summary": "Feedback",
        "operationId": "feedback_v1_decision_models__identifier__feedback_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Identifier"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Feedback"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/decision-models/{identifier}/workspace": {
      "get": {
        "summary": "Workspace",
        "operationId": "workspace_v1_decision_models__identifier__workspace_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/decision-models/{identifier}/simulations": {
      "post": {
        "summary": "Simulate",
        "operationId": "simulate_v1_decision_models__identifier__simulations_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Identifier"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/decision-models/{identifier}/judgments": {
      "post": {
        "summary": "Judge",
        "operationId": "judge_v1_decision_models__identifier__judgments_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Identifier"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JudgmentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/evidence/claims": {
      "post": {
        "summary": "Claim",
        "operationId": "claim_v1_evidence_claims_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Claim"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/graph/nodes": {
      "post": {
        "summary": "Node",
        "operationId": "node_v1_graph_nodes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraphNode"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/graph/edges": {
      "post": {
        "summary": "Edge",
        "operationId": "edge_v1_graph_edges_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraphEdge"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BalanceSheet": {
        "properties": {
          "assets": {
            "type": "number",
            "minimum": 0.0,
            "title": "Assets"
          },
          "liabilities": {
            "type": "number",
            "minimum": 0.0,
            "title": "Liabilities"
          },
          "equity": {
            "type": "number",
            "title": "Equity"
          },
          "tolerance": {
            "type": "number",
            "minimum": 0.0,
            "title": "Tolerance",
            "default": 0.01
          }
        },
        "type": "object",
        "required": [
          "assets",
          "liabilities",
          "equity"
        ],
        "title": "BalanceSheet"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PolicyRequest": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Question"
          },
          "context": {
            "$ref": "#/components/schemas/Context"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/praxis__services__policy_comparison__Option"
            },
            "type": "array",
            "maxItems": 20,
            "minItems": 2,
            "title": "Options"
          },
          "weights": {
            "additionalProperties": {
              "type": "number"
            },
            "propertyNames": {
              "enum": [
                "economic",
                "institutional",
                "diplomatic",
                "international_relations",
                "social",
                "fiscal",
                "environmental",
                "security",
                "implementation"
              ]
            },
            "type": "object",
            "maxProperties": 9,
            "minProperties": 1,
            "title": "Weights"
          },
          "scenarios": {
            "items": {
              "$ref": "#/components/schemas/praxis__services__policy_comparison__Scenario"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Scenarios"
          },
          "uncertainty_penalty": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Uncertainty Penalty",
            "default": 0.2
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "question",
          "context",
          "options",
          "weights"
        ],
        "title": "PolicyRequest"
      },
...
      "ValuationInputs": {
        "properties": {
          "cash_flows": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Cash Flows"
          },
          "discount_rate": {
            "type": "number",
            "maximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Discount Rate"
          },
          "terminal_growth": {
            "type": "number",
            "maximum": 1.0,
            "minimum": -1.0,
            "title": "Terminal Growth",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "cash_flows",
          "discount_rate"
        ],
        "title": "ValuationInputs"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}